+++ app/(routes)/user/Friends.tsx
