+++ app/components/screenComponents/FriendCard.tsx
