+++ app/lib/splitAmount.ts
